ceph: split osd code in files
Makes the code more readable and decoupled into files.
