+++ pkg/operator/ceph/cluster/osd/labels.go
@@ -1,0 +1,53 @@
+/*
+Copyright 2020 The Rook Authors. All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package osd
+
+import (
+	"fmt"
+	"strconv"
+
+	"github.com/rook/rook/pkg/operator/k8sutil"
+)
+
+const (
+	// CephDeviceSetLabelKey is the Rook device set label key
+	CephDeviceSetLabelKey = "ceph.rook.io/DeviceSet"
+	// CephSetIndexLabelKey is the Rook label key index
+	CephSetIndexLabelKey = "ceph.rook.io/setIndex"
+	// CephDeviceSetPVCIDLabelKey is the Rook PVC ID label key
+	CephDeviceSetPVCIDLabelKey = "ceph.rook.io/DeviceSetPVCId"
+	// OSDOverPVCLabelKey is the Rook PVC label key
+	OSDOverPVCLabelKey = "ceph.rook.io/pvc"
+)
+
+func makeStorageClassDeviceSetPVCLabel(storageClassDeviceSetName, pvcStorageClassDeviceSetPVCId string, setIndex int) map[string]string {
+	return map[string]string{
+		CephDeviceSetLabelKey:      storageClassDeviceSetName,
+		CephSetIndexLabelKey:       fmt.Sprintf("%d", setIndex),
+		CephDeviceSetPVCIDLabelKey: pvcStorageClassDeviceSetPVCId,
+	}
+}
+
+func (c *Cluster) getOSDLabels(osdID int, failureDomainValue string, portable bool) map[string]string {
+	return map[string]string{
+		k8sutil.AppAttr:     AppName,
+		k8sutil.ClusterAttr: c.clusterInfo.Namespace,
+		OsdIdLabelKey:       fmt.Sprintf("%d", osdID),
+		FailureDomainKey:    failureDomainValue,
+		portableKey:         strconv.FormatBool(portable),
+	}
+}
