+++ pkg/operator/ceph/cluster/osd/spec.go
@@ -18,11 +18,9 @@ limitations under the License.
 package osd
 
 import (
-	"encoding/json"
 	"fmt"
 	"path"
 	"strconv"
-	"strings"
 
 	"github.com/pkg/errors"
 	cephv1 "github.com/rook/rook/pkg/apis/ceph.rook.io/v1"
@@ -32,24 +30,11 @@ import (
 	"github.com/rook/rook/pkg/operator/ceph/controller"
 	"github.com/rook/rook/pkg/operator/k8sutil"
 	apps "k8s.io/api/apps/v1"
-	batch "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const (
-	osdStoreEnvVarName                  = "ROOK_OSD_STORE"
-	osdDatabaseSizeEnvVarName           = "ROOK_OSD_DATABASE_SIZE"
-	osdWalSizeEnvVarName                = "ROOK_OSD_WAL_SIZE"
-	osdJournalSizeEnvVarName            = "ROOK_OSD_JOURNAL_SIZE"
-	osdsPerDeviceEnvVarName             = "ROOK_OSDS_PER_DEVICE"
-	encryptedDeviceEnvVarName           = "ROOK_ENCRYPTED_DEVICE"
-	osdMetadataDeviceEnvVarName         = "ROOK_METADATA_DEVICE"
-	pvcBackedOSDVarName                 = "ROOK_PVC_BACKED_OSD"
-	blockPathVarName                    = "ROOK_BLOCK_PATH"
-	cvModeVarName                       = "ROOK_CV_MODE"
-	lvBackedPVVarName                   = "ROOK_LV_BACKED_PV"
-	CrushDeviceClassVarName             = "ROOK_OSD_CRUSH_DEVICE_CLASS"
 	rookBinariesMountPath               = "/rook"
 	rookBinariesVolumeName              = "rook-binaries"
 	activateOSDVolumeName               = "activate-osd"
@@ -58,16 +43,6 @@ const (
 	blockPVCMetadataMapperInitContainer = "blkdevmapper-metadata"
 	activatePVCOSDInitContainer         = "activate"
 	expandPVCOSDInitContainer           = "expand-bluefs"
-	// CephDeviceSetLabelKey is the Rook device set label key
-	CephDeviceSetLabelKey = "ceph.rook.io/DeviceSet"
-	// CephSetIndexLabelKey is the Rook label key index
-	CephSetIndexLabelKey = "ceph.rook.io/setIndex"
-	// CephDeviceSetPVCIDLabelKey is the Rook PVC ID label key
-	CephDeviceSetPVCIDLabelKey = "ceph.rook.io/DeviceSetPVCId"
-	// OSDOverPVCLabelKey is the Rook PVC label key
-	OSDOverPVCLabelKey = "ceph.rook.io/pvc"
-	udevPath           = "/run/udev"
-	udevVolName        = "run-udev"
 )
 
 const (
@@ -116,54 +91,6 @@ fi
 
 `
 )
-
-// OSDs on PVC using a certain storage class need to do some tuning
-const (
-	osdRecoverySleep = "0.1"
-	osdSnapTrimSleep = "2"
-	osdDeleteSleep   = "2"
-)
-
-func (c *Cluster) makeJob(osdProps osdProperties, provisionConfig *provisionConfig) (*batch.Job, error) {
-	podSpec, err := c.provisionPodTemplateSpec(osdProps, v1.RestartPolicyOnFailure, provisionConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	if !osdProps.onPVC() {
-		podSpec.Spec.NodeSelector = map[string]string{v1.LabelHostname: osdProps.crushHostname}
-	} else {
-		// This is not needed in raw mode and 14.2.8 brings it
-		// but we still want to do this not to lose backward compatibility with lvm based OSDs...
-		podSpec.Spec.InitContainers = append(podSpec.Spec.InitContainers, c.getPVCInitContainer(osdProps))
-		if osdProps.onPVCWithMetadata() {
-			podSpec.Spec.InitContainers = append(podSpec.Spec.InitContainers, c.getPVCMetadataInitContainer("/srv", osdProps))
-		}
-	}
-
-	job := &batch.Job{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      k8sutil.TruncateNodeName(prepareAppNameFmt, osdProps.crushHostname),
-			Namespace: c.clusterInfo.Namespace,
-			Labels: map[string]string{
-				k8sutil.AppAttr:     prepareAppName,
-				k8sutil.ClusterAttr: c.clusterInfo.Namespace,
-			},
-		},
-		Spec: batch.JobSpec{
-			Template: *podSpec,
-		},
-	}
-
-	if osdProps.onPVC() {
-		k8sutil.AddLabelToJob(OSDOverPVCLabelKey, osdProps.pvc.ClaimName, job)
-	}
-
-	k8sutil.AddRookVersionLabelToJob(job)
-	controller.AddCephVersionLabelToJob(c.clusterInfo.CephVersion, job)
-	k8sutil.SetOwnerRef(&job.ObjectMeta, &c.clusterInfo.OwnerRef)
-	return job, nil
-}
 
 func (c *Cluster) makeDeployment(osdProps osdProperties, osd OSDInfo, provisionConfig *provisionConfig) (*apps.Deployment, error) {
 	// If running on Octopus, we don't need to use the host PID namespace
@@ -551,84 +478,6 @@ func (c *Cluster) getActivateOSDInitContainer(osdID string, osdInfo OSDInfo, osd
 	return volume, container
 }
 
-func (c *Cluster) provisionPodTemplateSpec(osdProps osdProperties, restart v1.RestartPolicy, provisionConfig *provisionConfig) (*v1.PodTemplateSpec, error) {
-	copyBinariesVolume, copyBinariesContainer := c.getCopyBinariesContainer()
-
-	// ceph-volume is currently set up to use /etc/ceph/ceph.conf; this means no user config
-	// overrides will apply to ceph-volume, but this is unnecessary anyway
-	volumes := append(controller.PodVolumes(provisionConfig.DataPathMap, c.spec.DataDirHostPath, true), copyBinariesVolume)
-
-	// create a volume on /dev so the pod can access devices on the host
-	devVolume := v1.Volume{Name: "devices", VolumeSource: v1.VolumeSource{HostPath: &v1.HostPathVolumeSource{Path: "/dev"}}}
-	volumes = append(volumes, devVolume)
-	udevVolume := v1.Volume{Name: "udev", VolumeSource: v1.VolumeSource{HostPath: &v1.HostPathVolumeSource{Path: "/run/udev"}}}
-	volumes = append(volumes, udevVolume)
-
-	// If not running on PVC we mount the rootfs of the host to validate the presence of the LVM package
-	if !osdProps.onPVC() {
-		rootFSVolume := v1.Volume{Name: "rootfs", VolumeSource: v1.VolumeSource{HostPath: &v1.HostPathVolumeSource{Path: "/"}}}
-		volumes = append(volumes, rootFSVolume)
-	}
-
-	if osdProps.onPVC() {
-		// Create volume config for PVCs
-		volumes = append(volumes, getPVCOSDVolumes(&osdProps)...)
-	}
-
-	if len(volumes) == 0 {
-		return nil, errors.New("empty volumes")
-	}
-
-	provisionContainer, err := c.provisionOSDContainer(osdProps, copyBinariesContainer.VolumeMounts[0], provisionConfig)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to generate OSD provisioning container")
-	}
-
-	podSpec := v1.PodSpec{
-		ServiceAccountName: serviceAccountName,
-		InitContainers: []v1.Container{
-			*copyBinariesContainer,
-		},
-		Containers: []v1.Container{
-			provisionContainer,
-		},
-		RestartPolicy:     restart,
-		Volumes:           volumes,
-		HostNetwork:       c.spec.Network.IsHost(),
-		PriorityClassName: cephv1.GetOSDPriorityClassName(c.spec.PriorityClassNames),
-		SchedulerName:     osdProps.schedulerName,
-	}
-	if c.spec.Network.IsHost() {
-		podSpec.DNSPolicy = v1.DNSClusterFirstWithHostNet
-	}
-	if !osdProps.onPVC() {
-		cephv1.GetOSDPlacement(c.spec.Placement).ApplyToPodSpec(&podSpec)
-	} else {
-		osdProps.placement.ApplyToPodSpec(&podSpec)
-	}
-
-	podMeta := metav1.ObjectMeta{
-		Name: AppName,
-		Labels: map[string]string{
-			k8sutil.AppAttr:     prepareAppName,
-			k8sutil.ClusterAttr: c.clusterInfo.Namespace,
-			OSDOverPVCLabelKey:  osdProps.pvc.ClaimName,
-		},
-		Annotations: map[string]string{},
-	}
-
-	cephv1.GetOSDAnnotations(c.spec.Annotations).ApplyToObjectMeta(&podMeta)
-
-	// ceph-volume --dmcrypt uses cryptsetup that synchronizes with udev on
-	// host through semaphore
-	podSpec.HostIPC = osdProps.storeConfig.EncryptedDevice
-
-	return &v1.PodTemplateSpec{
-		ObjectMeta: podMeta,
-		Spec:       podSpec,
-	}, nil
-}
-
 // Currently we can't mount a block mode pv directly to a privileged container
 // So we mount it to a non privileged init container and then copy it to a common directory mounted inside init container
 // and the privileged provision container.
@@ -776,418 +625,4 @@ func (c *Cluster) getExpandPVCInitContainer(osdProps osdProperties, osdID string
 	}
 
 	return container
-}
-func (c *Cluster) getConfigEnvVars(osdProps osdProperties, dataDir string) []v1.EnvVar {
-	envVars := []v1.EnvVar{
-		nodeNameEnvVar(osdProps.crushHostname),
-		{Name: "ROOK_CLUSTER_ID", Value: string(c.clusterInfo.OwnerRef.UID)},
-		k8sutil.PodIPEnvVar(k8sutil.PrivateIPEnvVar),
-		k8sutil.PodIPEnvVar(k8sutil.PublicIPEnvVar),
-		opmon.PodNamespaceEnvVar(c.clusterInfo.Namespace),
-		opmon.EndpointEnvVar(),
-		opmon.SecretEnvVar(),
-		opmon.CephUsernameEnvVar(),
-		opmon.CephSecretEnvVar(),
-		k8sutil.ConfigDirEnvVar(dataDir),
-		k8sutil.ConfigOverrideEnvVar(),
-		{Name: "ROOK_FSID", ValueFrom: &v1.EnvVarSource{
-			SecretKeyRef: &v1.SecretKeySelector{
-				LocalObjectReference: v1.LocalObjectReference{Name: "rook-ceph-mon"},
-				Key:                  "fsid",
-			},
-		}},
-		k8sutil.NodeEnvVar(),
-	}
-
-	// Give a hint to the prepare pod for what the host in the CRUSH map should be
-	crushmapHostname := osdProps.crushHostname
-	if !osdProps.portable && osdProps.onPVC() {
-		// If it's a pvc that's not portable we only know what the host name should be when inside the osd prepare pod
-		crushmapHostname = ""
-	}
-	envVars = append(envVars, v1.EnvVar{Name: "ROOK_CRUSHMAP_HOSTNAME", Value: crushmapHostname})
-
-	// Append ceph-volume environment variables
-	envVars = append(envVars, cephVolumeEnvVar()...)
-
-	if osdProps.storeConfig.DatabaseSizeMB != 0 {
-		envVars = append(envVars, v1.EnvVar{Name: osdDatabaseSizeEnvVarName, Value: strconv.Itoa(osdProps.storeConfig.DatabaseSizeMB)})
-	}
-
-	if osdProps.storeConfig.WalSizeMB != 0 {
-		envVars = append(envVars, v1.EnvVar{Name: osdWalSizeEnvVarName, Value: strconv.Itoa(osdProps.storeConfig.WalSizeMB)})
-	}
-
-	if osdProps.storeConfig.OSDsPerDevice != 0 {
-		envVars = append(envVars, v1.EnvVar{Name: osdsPerDeviceEnvVarName, Value: strconv.Itoa(osdProps.storeConfig.OSDsPerDevice)})
-	}
-
-	if osdProps.storeConfig.EncryptedDevice {
-		envVars = append(envVars, v1.EnvVar{Name: encryptedDeviceEnvVarName, Value: "true"})
-	}
-
-	return envVars
-}
-
-func (c *Cluster) provisionOSDContainer(osdProps osdProperties, copyBinariesMount v1.VolumeMount, provisionConfig *provisionConfig) (v1.Container, error) {
-	envVars := c.getConfigEnvVars(osdProps, k8sutil.DataDir)
-
-	// enable debug logging in the prepare job
-	envVars = append(envVars, setDebugLogLevelEnvVar(true))
-
-	// Drive Groups cannot be used to configure OSDs on PVCs, so ignore if this is a PVC config
-	// This shouldn't ever happen, but do the PVC check to be sure
-	if len(osdProps.driveGroups) > 0 && !osdProps.onPVC() {
-		v, err := c.getDriveGroupEnvVar(osdProps)
-		if err != nil {
-			// Because OSD creation via drive groups should take precedent over other types of drive
-			// creation, if there is an error here, we should fail. Allowing OSD creation to proceed
-			// without drive group information could result in OSD configs being created which the
-			// user does not want.
-			return v1.Container{}, errors.Wrapf(err, "failed to get drive group info for OSD provisioning container")
-		}
-		// An env var with no name means there are no groups, don't add the var
-		if v.Name != "" {
-			envVars = append(envVars, v)
-		}
-	}
-
-	// only 1 of device list, device filter, device path filter and use all devices can be specified.  We prioritize in that order.
-	if len(osdProps.devices) > 0 {
-		configuredDevices := []config.ConfiguredDevice{}
-		for _, device := range osdProps.devices {
-			id := device.Name
-			if device.FullPath != "" {
-				id = device.FullPath
-			}
-			cd := config.ConfiguredDevice{
-				ID:          id,
-				StoreConfig: config.ToStoreConfig(device.Config),
-			}
-			configuredDevices = append(configuredDevices, cd)
-		}
-		marshalledDevices, err := json.Marshal(configuredDevices)
-		if err != nil {
-			return v1.Container{}, errors.Wrapf(err, "failed to JSON marshal configured devices for node %q", osdProps.crushHostname)
-		}
-		envVars = append(envVars, dataDevicesEnvVar(string(marshalledDevices)))
-	} else if osdProps.selection.DeviceFilter != "" {
-		envVars = append(envVars, deviceFilterEnvVar(osdProps.selection.DeviceFilter))
-	} else if osdProps.selection.DevicePathFilter != "" {
-		envVars = append(envVars, devicePathFilterEnvVar(osdProps.selection.DevicePathFilter))
-	} else if osdProps.selection.GetUseAllDevices() {
-		envVars = append(envVars, deviceFilterEnvVar("all"))
-	}
-	envVars = append(envVars, v1.EnvVar{Name: "ROOK_CEPH_VERSION", Value: c.clusterInfo.CephVersion.CephVersionFormatted()})
-
-	if osdProps.metadataDevice != "" {
-		envVars = append(envVars, metadataDeviceEnvVar(osdProps.metadataDevice))
-	}
-
-	volumeMounts := append(controller.CephVolumeMounts(provisionConfig.DataPathMap, true), []v1.VolumeMount{
-		{Name: "devices", MountPath: "/dev"},
-		{Name: "udev", MountPath: "/run/udev"},
-		copyBinariesMount,
-	}...)
-
-	// If not running on PVC we mount the rootfs of the host to validate the presence of the LVM package
-	if !osdProps.onPVC() {
-		volumeMounts = append(volumeMounts, v1.VolumeMount{Name: "rootfs", MountPath: "/rootfs", ReadOnly: true})
-	}
-
-	// If the OSD runs on PVC
-	if osdProps.onPVC() {
-		volumeMounts = append(volumeMounts, getPvcOSDBridgeMount(osdProps.pvc.ClaimName))
-		// The device list is read by the Rook CLI via environment variables so let's add them
-		configuredDevices := []config.ConfiguredDevice{
-			{
-				ID:          fmt.Sprintf("/mnt/%s", osdProps.pvc.ClaimName),
-				StoreConfig: config.NewStoreConfig(),
-			},
-		}
-		if osdProps.onPVCWithMetadata() {
-			volumeMounts = append(volumeMounts, getPvcMetadataOSDBridgeMount(osdProps.metadataPVC.ClaimName))
-			configuredDevices = append(configuredDevices,
-				config.ConfiguredDevice{
-					ID:          fmt.Sprintf("/srv/%s", osdProps.metadataPVC.ClaimName),
-					StoreConfig: config.NewStoreConfig(),
-				})
-		}
-		marshalledDevices, err := json.Marshal(configuredDevices)
-		if err != nil {
-			return v1.Container{}, errors.Wrapf(err, "failed to JSON marshal configured devices for PVC %q", osdProps.crushHostname)
-		}
-		envVars = append(envVars, dataDevicesEnvVar(string(marshalledDevices)))
-		envVars = append(envVars, pvcBackedOSDEnvVar("true"))
-		envVars = append(envVars, crushDeviceClassEnvVar(osdProps.crushDeviceClass))
-	}
-
-	// run privileged always since we always mount /dev
-	privileged := true
-	runAsUser := int64(0)
-	runAsNonRoot := false
-	readOnlyRootFilesystem := false
-
-	osdProvisionContainer := v1.Container{
-		Command:      []string{path.Join(rookBinariesMountPath, "tini")},
-		Args:         []string{"--", path.Join(rookBinariesMountPath, "rook"), "ceph", "osd", "provision"},
-		Name:         "provision",
-		Image:        c.spec.CephVersion.Image,
-		VolumeMounts: volumeMounts,
-		Env:          envVars,
-		SecurityContext: &v1.SecurityContext{
-			Privileged:             &privileged,
-			RunAsUser:              &runAsUser,
-			RunAsNonRoot:           &runAsNonRoot,
-			ReadOnlyRootFilesystem: &readOnlyRootFilesystem,
-		},
-		Resources: cephv1.GetPrepareOSDResources(c.spec.Resources),
-	}
-
-	return osdProvisionContainer, nil
-}
-
-func getPvcOSDBridgeMount(claimName string) v1.VolumeMount {
-	return v1.VolumeMount{Name: fmt.Sprintf("%s-bridge", claimName), MountPath: "/mnt"}
-}
-
-func getPvcOSDBridgeMountActivate(mountPath, claimName string) v1.VolumeMount {
-	return v1.VolumeMount{Name: fmt.Sprintf("%s-bridge", claimName), MountPath: mountPath, SubPath: path.Base(mountPath)}
-}
-
-func getPvcMetadataOSDBridgeMount(claimName string) v1.VolumeMount {
-	return v1.VolumeMount{Name: fmt.Sprintf("%s-bridge", claimName), MountPath: "/srv"}
-}
-
-func (c *Cluster) skipVolumeForDirectory(path string) bool {
-	// If attempting to add a directory at /var/lib/rook, we need to skip the volume and volume mount
-	// since the dataDirHostPath is always mounting at /var/lib/rook
-	return path == k8sutil.DataDir
-}
-
-func getPVCOSDVolumes(osdProps *osdProperties) []v1.Volume {
-	volumes := []v1.Volume{
-		{
-			Name: osdProps.pvc.ClaimName,
-			VolumeSource: v1.VolumeSource{
-				PersistentVolumeClaim: &osdProps.pvc,
-			},
-		},
-		{
-			// We need a bridge mount which is basically a common volume mount between the non privileged init container
-			// and the privileged provision container or osd daemon container
-			// The reason for this is mentioned in the comment for getPVCInitContainer() method
-			Name: fmt.Sprintf("%s-bridge", osdProps.pvc.ClaimName),
-			VolumeSource: v1.VolumeSource{
-				EmptyDir: &v1.EmptyDirVolumeSource{
-					Medium: "Memory",
-				},
-			},
-		},
-	}
-
-	// If we have a metadata PVC let's add it
-	if osdProps.onPVCWithMetadata() {
-		metadataPVCVolume := []v1.Volume{
-			{
-				Name: osdProps.metadataPVC.ClaimName,
-				VolumeSource: v1.VolumeSource{
-					PersistentVolumeClaim: &osdProps.metadataPVC,
-				},
-			},
-			{
-				// We need a bridge mount which is basically a common volume mount between the non privileged init container
-				// and the privileged provision container or osd daemon container
-				// The reason for this is mentioned in the comment for getPVCInitContainer() method
-				Name: fmt.Sprintf("%s-bridge", osdProps.metadataPVC.ClaimName),
-				VolumeSource: v1.VolumeSource{
-					EmptyDir: &v1.EmptyDirVolumeSource{
-						Medium: "Memory",
-					},
-				},
-			},
-		}
-
-		volumes = append(volumes, metadataPVCVolume...)
-	}
-
-	logger.Debugf("volumes are %+v", volumes)
-
-	return volumes
-}
-
-func (c *Cluster) getDriveGroupEnvVar(osdProps osdProperties) (v1.EnvVar, error) {
-	if len(osdProps.driveGroups) == 0 {
-		return v1.EnvVar{}, nil
-	}
-
-	b, err := MarshalAsDriveGroupBlobs(osdProps.driveGroups)
-	if err != nil {
-		return v1.EnvVar{}, errors.Wrap(err, "failed to marshal drive groups into an env var")
-	}
-	return driveGroupsEnvVar(b), nil
-}
-
-func nodeNameEnvVar(name string) v1.EnvVar {
-	return v1.EnvVar{Name: "ROOK_NODE_NAME", Value: name}
-}
-
-func driveGroupsEnvVar(driveGroups string) v1.EnvVar {
-	return v1.EnvVar{Name: "ROOK_DRIVE_GROUPS", Value: driveGroups}
-}
-
-func dataDevicesEnvVar(dataDevices string) v1.EnvVar {
-	return v1.EnvVar{Name: "ROOK_DATA_DEVICES", Value: dataDevices}
-}
-
-func deviceFilterEnvVar(filter string) v1.EnvVar {
-	return v1.EnvVar{Name: "ROOK_DATA_DEVICE_FILTER", Value: filter}
-}
-
-func devicePathFilterEnvVar(filter string) v1.EnvVar {
-	return v1.EnvVar{Name: "ROOK_DATA_DEVICE_PATH_FILTER", Value: filter}
-}
-
-func metadataDeviceEnvVar(metadataDevice string) v1.EnvVar {
-	return v1.EnvVar{Name: osdMetadataDeviceEnvVarName, Value: metadataDevice}
-}
-
-func pvcBackedOSDEnvVar(pvcBacked string) v1.EnvVar {
-	return v1.EnvVar{Name: pvcBackedOSDVarName, Value: pvcBacked}
-}
-
-func setDebugLogLevelEnvVar(debug bool) v1.EnvVar {
-	level := "INFO"
-	if debug {
-		level = "DEBUG"
-	}
-	return v1.EnvVar{Name: "ROOK_LOG_LEVEL", Value: level}
-}
-
-func blockPathEnvVariable(lvPath string) v1.EnvVar {
-	return v1.EnvVar{Name: blockPathVarName, Value: lvPath}
-}
-
-func cvModeEnvVariable(cvMode string) v1.EnvVar {
-	return v1.EnvVar{Name: cvModeVarName, Value: cvMode}
-}
-
-func lvBackedPVEnvVar(lvBackedPV string) v1.EnvVar {
-	return v1.EnvVar{Name: lvBackedPVVarName, Value: lvBackedPV}
-}
-
-func crushDeviceClassEnvVar(crushDeviceClass string) v1.EnvVar {
-	return v1.EnvVar{Name: CrushDeviceClassVarName, Value: crushDeviceClass}
-}
-
-func osdOnSDNFlag(network cephv1.NetworkSpec) []string {
-	var args []string
-	// OSD fails to find the right IP to bind to when running on SDN
-	// for more details: https://github.com/rook/rook/issues/3140
-	if !network.IsHost() {
-		args = append(args, "--ms-learn-addr-from-peer=false")
-	}
-
-	return args
-}
-
-func makeStorageClassDeviceSetPVCID(storageClassDeviceSetName string, setIndex int) (pvcID, pvcLabelSelector string) {
-	pvcStorageClassDeviceSetPVCId := fmt.Sprintf("%s-%d", storageClassDeviceSetName, setIndex)
-	return pvcStorageClassDeviceSetPVCId, fmt.Sprintf("%s=%s", CephDeviceSetPVCIDLabelKey, pvcStorageClassDeviceSetPVCId)
-}
-
-// This is the new function that generates the labels
-// It includes the pvcTemplateName in it
-func makeStorageClassDeviceSetPVCIDNew(storageClassDeviceSetName, pvcTemplateName string, setIndex int) (pvcID, pvcLabelSelector string) {
-	pvcStorageClassDeviceSetPVCId := fmt.Sprintf("%s-%s-%d", storageClassDeviceSetName, strings.Replace(pvcTemplateName, " ", "-", -1), setIndex)
-	return pvcStorageClassDeviceSetPVCId, fmt.Sprintf("%s=%s", CephDeviceSetPVCIDLabelKey, pvcStorageClassDeviceSetPVCId)
-}
-
-func makeStorageClassDeviceSetPVCLabel(storageClassDeviceSetName, pvcStorageClassDeviceSetPVCId string, setIndex int) map[string]string {
-	return map[string]string{
-		CephDeviceSetLabelKey:      storageClassDeviceSetName,
-		CephSetIndexLabelKey:       fmt.Sprintf("%d", setIndex),
-		CephDeviceSetPVCIDLabelKey: pvcStorageClassDeviceSetPVCId,
-	}
-}
-
-func (c *Cluster) getOSDLabels(osdID int, failureDomainValue string, portable bool) map[string]string {
-	return map[string]string{
-		k8sutil.AppAttr:     AppName,
-		k8sutil.ClusterAttr: c.clusterInfo.Namespace,
-		OsdIdLabelKey:       fmt.Sprintf("%d", osdID),
-		FailureDomainKey:    failureDomainValue,
-		portableKey:         strconv.FormatBool(portable),
-	}
-}
-
-func cephVolumeEnvVar() []v1.EnvVar {
-	return []v1.EnvVar{
-		{Name: "CEPH_VOLUME_DEBUG", Value: "1"},
-		{Name: "CEPH_VOLUME_SKIP_RESTORECON", Value: "1"},
-		// LVM will avoid interaction with udev.
-		// LVM will manage the relevant nodes in /dev directly.
-		{Name: "DM_DISABLE_UDEV", Value: "1"},
-	}
-}
-
-func osdActivateEnvVar() []v1.EnvVar {
-	monEnvVars := []v1.EnvVar{
-		{Name: "ROOK_CEPH_MON_HOST",
-			ValueFrom: &v1.EnvVarSource{
-				SecretKeyRef: &v1.SecretKeySelector{LocalObjectReference: v1.LocalObjectReference{
-					Name: "rook-ceph-config"},
-					Key: "mon_host"}}},
-		{Name: "CEPH_ARGS", Value: "-m $(ROOK_CEPH_MON_HOST)"},
-	}
-
-	return append(cephVolumeEnvVar(), monEnvVars...)
-}
-
-func (c *Cluster) osdRunFlagTuningOnPVC(osdID int) error {
-	who := fmt.Sprintf("osd.%d", osdID)
-	do := make(map[string]string)
-
-	// Time in seconds to sleep before next recovery or backfill op
-	do["osd_recovery_sleep"] = osdRecoverySleep
-	// Time in seconds to sleep before next snap trim
-	do["osd_snap_trim_sleep"] = osdSnapTrimSleep
-	// Time in seconds to sleep before next removal transaction
-	do["osd_delete_sleep"] = osdDeleteSleep
-
-	monStore := opconfig.GetMonStore(c.context, c.clusterInfo)
-
-	for flag, val := range do {
-		err := monStore.Set(who, flag, val)
-		if err != nil {
-			return errors.Wrapf(err, "failed to set %q to %q on %q", flag, val, who)
-		}
-	}
-
-	return nil
-}
-
-func getUdevVolume() (v1.Volume, v1.VolumeMount) {
-	volume := v1.Volume{
-		Name: udevVolName,
-		VolumeSource: v1.VolumeSource{
-			HostPath: &v1.HostPathVolumeSource{Path: udevPath},
-		},
-	}
-
-	volumeMounts := v1.VolumeMount{
-		Name:      udevVolName,
-		MountPath: udevPath,
-	}
-
-	return volume, volumeMounts
-}
-
-// PrivilegedContext returns a privileged Pod security context
-func PrivilegedContext() *v1.SecurityContext {
-	privileged := true
-
-	return &v1.SecurityContext{
-		Privileged: &privileged,
-	}
 }
