+++ pkg/operator/ceph/cluster/osd/spec_test.go
@@ -359,17 +359,6 @@ func TestHostNetwork(t *testing.T) {
 	assert.Equal(t, v1.DNSClusterFirstWithHostNet, r.Spec.Template.Spec.DNSPolicy)
 }
 
-func TestOsdOnSDNFlag(t *testing.T) {
-	network := cephv1.NetworkSpec{}
-
-	args := osdOnSDNFlag(network)
-	assert.NotEmpty(t, args)
-
-	network.Provider = "host"
-	args = osdOnSDNFlag(network)
-	assert.Empty(t, args)
-}
-
 func TestOsdPrepareResources(t *testing.T) {
 	clientset := fake.NewSimpleClientset()
 
@@ -393,102 +382,4 @@ func TestOsdPrepareResources(t *testing.T) {
 	assert.Equal(t, "0", r.Requests.Cpu().String())
 	assert.Equal(t, "0", r.Limits.Memory().String())
 	assert.Equal(t, "250", r.Requests.Memory().String())
-}
-
-func TestCephVolumeEnvVar(t *testing.T) {
-	cvEnv := cephVolumeEnvVar()
-	assert.Equal(t, "CEPH_VOLUME_DEBUG", cvEnv[0].Name)
-	assert.Equal(t, "1", cvEnv[0].Value)
-	assert.Equal(t, "CEPH_VOLUME_SKIP_RESTORECON", cvEnv[1].Name)
-	assert.Equal(t, "1", cvEnv[1].Value)
-	assert.Equal(t, "DM_DISABLE_UDEV", cvEnv[2].Name)
-	assert.Equal(t, "1", cvEnv[1].Value)
-}
-
-func TestOsdActivateEnvVar(t *testing.T) {
-	osdActivateEnv := osdActivateEnvVar()
-	assert.Equal(t, 5, len(osdActivateEnv))
-	assert.Equal(t, "CEPH_VOLUME_DEBUG", osdActivateEnv[0].Name)
-	assert.Equal(t, "1", osdActivateEnv[0].Value)
-	assert.Equal(t, "CEPH_VOLUME_SKIP_RESTORECON", osdActivateEnv[1].Name)
-	assert.Equal(t, "1", osdActivateEnv[1].Value)
-	assert.Equal(t, "DM_DISABLE_UDEV", osdActivateEnv[2].Name)
-	assert.Equal(t, "1", osdActivateEnv[1].Value)
-	assert.Equal(t, "ROOK_CEPH_MON_HOST", osdActivateEnv[3].Name)
-	assert.Equal(t, "CEPH_ARGS", osdActivateEnv[4].Name)
-	assert.Equal(t, "-m $(ROOK_CEPH_MON_HOST)", osdActivateEnv[4].Value)
-}
-
-func TestDriveGroups(t *testing.T) {
-	clientset := fake.NewSimpleClientset()
-	context := &clusterd.Context{Clientset: clientset, ConfigDir: "/var/lib/rook", Executor: &exectest.MockExecutor{}}
-
-	dataPathMap := &provisionConfig{
-		DataPathMap: opconfig.NewDatalessDaemonDataPathMap("ns", "/var/lib/rook"),
-	}
-
-	driveGroups := cephv1.DriveGroupsSpec{
-		cephv1.DriveGroup{
-			Name: "group1",
-			Spec: cephv1.DriveGroupSpec{
-				"data_devices": map[string]string{
-					"all": "true",
-				},
-			},
-			Placement: rookv1.Placement{},
-		},
-	}
-	clusterInfo := &cephclient.ClusterInfo{Namespace: "ns"}
-	spec := cephv1.ClusterSpec{
-		DataDirHostPath: "/var/lib/rook",
-		CephVersion:     cephv1.CephVersionSpec{Image: "test-image"},
-		DriveGroups:     driveGroups,
-		Storage: rookv1.StorageScopeSpec{
-			UseAllNodes: true,
-			Selection:   rookv1.Selection{},
-			Nodes:       []rookv1.Node{{Name: "node1"}, {Name: "node2"}},
-		},
-	}
-	c := New(context, clusterInfo, spec, "rook/rook:myversion")
-
-	n := c.spec.Storage.ResolveNode("node1")
-
-	osdProp := osdProperties{
-		crushHostname: n.Name,
-	}
-
-	assertNoDriveGroups := func(e []v1.EnvVar) {
-		for _, v := range e {
-			if v.Name == "ROOK_DRIVE_GROUPS" {
-				t.Helper()
-				t.Errorf("ROOK_DRIVE_GROUPS should not be specified to provisioning pod but is: %+v", v)
-			}
-		}
-	}
-
-	assertDriveGroups := func(e []v1.EnvVar) {
-		for _, v := range e {
-			if v.Name == "ROOK_DRIVE_GROUPS" {
-				return
-			}
-		}
-		t.Helper()
-		t.Errorf("ROOK_DRIVE_GROUPS should be specified to provisioning pod but is not: %+v", e)
-	}
-
-	// Drive groups are not specified; should NOT have env var specified
-	clusterInfo.CephVersion = cephver.CephVersion{Major: 15, Minor: 2, Extra: 5}
-	spec.DriveGroups = cephv1.DriveGroupsSpec{}
-	job, err := c.makeJob(osdProp, dataPathMap)
-	assert.NotNil(t, job)
-	assert.NoError(t, err)
-	assertNoDriveGroups(job.Spec.Template.Spec.Containers[0].Env)
-
-	// Drive groups are specified; should have env var specified
-	osdProp.driveGroups = driveGroups
-	spec.DriveGroups = driveGroups
-	job, err = c.makeJob(osdProp, dataPathMap)
-	assert.NotNil(t, job)
-	assert.NoError(t, err)
-	assertDriveGroups(job.Spec.Template.Spec.Containers[0].Env)
 }
